Dédoublonne les utils trophées après merge de la page trophées
Consolide utils/trophee.tsx dans utils/trophies.tsx (source unique) :
- trophyImageUrl / trophyMedalColor : une seule copie
- le hook getUserTrophiesData réutilise fetchUserTrophies
- fusionne TrophyUserData dans Trophy (level: number)
